+++ pkg/api/clusters/cluster/cluster.go
@@ -2,7 +2,12 @@ package cluster
 
 import (
 	"context"
+	"regexp"
+	"strings"
 	"time"
+
+	kobsClientsetVersioned "github.com/kobsio/kobs/pkg/generated/clientset/versioned"
+	"github.com/kobsio/kobs/pkg/generated/proto"
 
 	"github.com/sirupsen/logrus"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -11,15 +16,17 @@ import (
 )
 
 var (
-	log = logrus.WithFields(logrus.Fields{"package": "cluster"})
+	log       = logrus.WithFields(logrus.Fields{"package": "cluster"})
+	slugifyRe = regexp.MustCompile("[^a-z0-9]+")
 )
 
 // Cluster is a Kubernetes cluster. It contains all required fields to interact with the cluster and it's services.
 type Cluster struct {
-	cache     Cache
-	clientset *kubernetes.Clientset
-	options   Options
-	name      string
+	cache         Cache
+	clientset     *kubernetes.Clientset
+	kobsClientset *kobsClientsetVersioned.Clientset
+	options       Options
+	name          string
 }
 
 // Options contains various options, which could be set for a cluster. For example a user can set the cache duration for
@@ -101,18 +108,70 @@ func (c *Cluster) GetResources(ctx context.Context, namespace, path, resource, p
 	return string(res), nil
 }
 
+// GetApplications returns a list of applications gor the given namespace. It also adds the cluster, namespace and
+// application name to the application CR, so that this information must not be specified by the user in the CR.
+func (c *Cluster) GetApplications(ctx context.Context, namespace string) ([]*proto.Application, error) {
+	applicationsList, err := c.kobsClientset.KobsV1alpha1().Applications(namespace).List(ctx, metav1.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	var applications []*proto.Application
+
+	for _, app := range applicationsList.Items {
+		if app.Spec == nil {
+			continue
+		}
+
+		application := app.Spec
+		application.Cluster = c.name
+		application.Namespace = app.Namespace
+		application.Name = app.Name
+
+		applications = append(applications, application)
+	}
+
+	return applications, nil
+}
+
+// GetApplication returns a application for the given namespace and name. After the application is retrieved we replace,
+// the cluster, namespace and name in the spec of the Application CR. This is needed, so that the user doesn't have to,
+// provide these fields.
+func (c *Cluster) GetApplication(ctx context.Context, namespace, name string) (*proto.Application, error) {
+	applicationCR, err := c.kobsClientset.KobsV1alpha1().Applications(namespace).Get(ctx, name, metav1.GetOptions{})
+	if err != nil {
+		return nil, err
+	}
+
+	application := applicationCR.Spec
+	application.Cluster = c.name
+	application.Namespace = namespace
+	application.Name = name
+
+	return application, nil
+}
+
 // NewCluster returns a new cluster. Each cluster must have a unique name and a client to make requests against the
 // Kubernetes API server of this cluster.
 func NewCluster(name string, restConfig *rest.Config) (*Cluster, error) {
 	clientset, err := kubernetes.NewForConfig(restConfig)
 	if err != nil {
-		log.WithError(err).Debugf("Could not create clientset.")
+		log.WithError(err).Debugf("Could not create Kubernetes clientset.")
 		return nil, err
 	}
 
+	kobsClientset, err := kobsClientsetVersioned.NewForConfig(restConfig)
+	if err != nil {
+		log.WithError(err).Debugf("Could not create kobs clientset.")
+		return nil, err
+	}
+
+	name = strings.Trim(slugifyRe.ReplaceAllString(strings.ToLower(name), "-"), "-")
+
 	return &Cluster{
-		clientset: clientset,
-		name:      name,
+		clientset:     clientset,
+		kobsClientset: kobsClientset,
+		name:          name,
 	}, nil
 }